chore: refactor getCreateOperationForProfile to use getCreatePayloadFromDocumentModel
