+++ packages/wallet/src/functions/getCreateOperationForProfile.ts
@@ -11,12 +11,10 @@
  * limitations under the License.
  */
 
-import canonicalize from 'canonicalize';
-import base64url from 'base64url';
-import { canonicalizeThenHashThenEncode } from './sidetreeEncoding';
+import { PublicKeyPurpose } from '@sidetree/common';
 import { toKeyPair } from './toKeyPair';
 import { SidetreeCreateOperation, SidetreeReplaceOptions } from '../types';
-import { PublicKeyPurpose } from '@sidetree/common';
+import { getCreatePayloadFromDocumentModel } from './getCreatePayloadFromDocumentModel';
 
 export const getCreateOperationForProfile = async (
   mnemonic: string,
@@ -29,56 +27,34 @@ export const getCreateOperationForProfile = async (
   }
   const signingKeyPair = await toKeyPair(mnemonic, index, 'Ed25519');
   const keyAgreementKeyPair = await toKeyPair(mnemonic, index, 'X25519');
-
-  const delta_object = {
-    update_commitment: canonicalizeThenHashThenEncode(
-      signingKeyPair.publicKeyJwk
-    ),
-    patches: [
+  const documentModel = {
+    public_keys: [
       {
-        action: 'replace',
-        document: {
-          public_keys: [
-            {
-              id: signingKeyPair.id.split('#').pop(),
-              type: 'JsonWebKey2020',
-              jwk: signingKeyPair.publicKeyJwk,
-              purpose: [
-                PublicKeyPurpose.General,
-                PublicKeyPurpose.Auth,
-                PublicKeyPurpose.AssertionMethod,
-                PublicKeyPurpose.CapabilityInvocation,
-                PublicKeyPurpose.CapabilityDelegation,
-              ],
-            },
-            {
-              id: keyAgreementKeyPair.id.split('#').pop(),
-              type: 'JsonWebKey2020',
-              jwk: keyAgreementKeyPair.publicKeyJwk,
-              purpose: [
-                PublicKeyPurpose.General,
-                PublicKeyPurpose.KeyAgreement,
-              ],
-            },
-          ],
-          ...options,
-        },
+        id: signingKeyPair.id.split('#').pop(),
+        type: 'JsonWebKey2020',
+        jwk: signingKeyPair.publicKeyJwk,
+        purpose: [
+          PublicKeyPurpose.General,
+          PublicKeyPurpose.Auth,
+          PublicKeyPurpose.AssertionMethod,
+          PublicKeyPurpose.CapabilityInvocation,
+          PublicKeyPurpose.CapabilityDelegation,
+        ],
+      },
+      {
+        id: keyAgreementKeyPair.id.split('#').pop(),
+        type: 'JsonWebKey2020',
+        jwk: keyAgreementKeyPair.publicKeyJwk,
+        purpose: [PublicKeyPurpose.General, PublicKeyPurpose.KeyAgreement],
       },
     ],
+    ...options,
   };
-  const delta = base64url.encode(canonicalize(delta_object));
-  const canonical_suffix_data = canonicalize({
-    delta_hash: canonicalizeThenHashThenEncode(delta_object),
-    recovery_commitment: canonicalizeThenHashThenEncode(
-      signingKeyPair.publicKeyJwk
-    ),
-  });
-  const suffix_data = base64url.encode(canonical_suffix_data);
-  const createOperation: SidetreeCreateOperation = {
-    type: 'create',
-    suffix_data,
-    delta,
-  };
+  const createOperation = await getCreatePayloadFromDocumentModel(
+    documentModel,
+    signingKeyPair.publicKeyJwk,
+    signingKeyPair.publicKeyJwk
+  );
 
   return createOperation;
 };
